requirements.txt
# Web Scraping & HTML Processing
beautifulsoup4>=4.12.0
selenium>=4.15.0
requests>=2.31.0
html2text>=2024.2.0
lxml>=4.9.0

# PDF Processing
PyPDF2>=3.0.0
pypdf>=4.0.0
reportlab>=4.0.0
python-pptx>=0.6.21

# LLM & Embeddings
openai>=1.30.0
anthropic>=0.28.0
sentence-transformers>=2.5.0
huggingface-hub>=0.23.0
transformers>=4.40.0
torch>=2.0.0,<3.0

# Vector Database
chromadb>=0.5.0
pinecone-client>=5.0.0

# Backend API
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database & ORM
sqlalchemy>=2.0.0
alembic>=1.13.0

# Data Processing & ML Evaluation
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
nltk>=3.8.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.67.0
loguru>=0.7.0
jsonlines>=4.0.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.13.0